==> analyses/0_exploration/1_kernel_exploration.Rmd <==
---
title: Kernal word exploration
date: "`r Sys.Date()`"
output:
  html_document:
    highlight: tango
    toc: true
    number_sections: false
    theme: cerulean
    toc_float: true
    code_folding: hide
---
Broad goal: (1) identify set of words on independent theoretical grounds (e.g. semantic networks) that seem core to vocabulary. (2) test these predictions on corpus of children's vocabulary where the prediction is that children who produce key words at *t1* will produce more words in general at *t2*.

```{r setup, message=FALSE, warning = F}
library(knitr)

opts_chunk$set(echo = T, message = F, warning = F, 
               error = F, cache = T, tidy = F)

library(tidyverse)
library(feather)

theme_set(theme_classic(base_size = 10))

```



```{r}
full_dict <- read_delim("allGKH.csv", delim = ";") %>%
  mutate_all(as.factor) %>%
  rename(word = WORD,
         pos = POS, 
         structure = STRUCTURE,
         dict = Dictionary)
```

Why are stucture categories not mutually exclusive? e.g. 
```{r}
full_dict %>%
  filter(dict == "Longman") %>%
  head()
```
because different senses.

Take min dist sense for each word. 

Let's look only at the two dictionaries that are intended for speakers with limited English knowledge.
```{r}
longman_dict <- full_dict %>%
  filter(dict == "Longman") %>%
  filter(structure == "0")

cambridge_dict <- full_dict %>%
  filter(dict == "Cambridge") %>%
  filter(structure == "0")

# intersection of core words in both dictionaries
core_words <- intersect(cambridge_dict$word,
                        longman_dict$word)

#write_feather(data.frame(word = core_words), "kernel_core_words.feather")
```

There are `r length(core_words)` words in the kernal of both dictionaries.
